// ScrabbleSharp.Frontend/scrabblesharp-frontend/src/api/grpc/ScrabbleSharp_pb.ts
export enum Multiplier {
  /**
   * @generated from enum value: NONE = 0;
   */
  NONE = 0,

  /**
   * @generated from enum value: DOUBLE_LETTER = 1;
   */
  DOUBLE_LETTER = 1,

  /**
   * @generated from enum value: TRIPLE_LETTER = 2;
   */
  TRIPLE_LETTER = 2,

  /**
   * @generated from enum value: QUADRUPLE_LETTER = 3;
   */
  QUADRUPLE_LETTER = 3,

  /**
   * @generated from enum value: DOUBLE_WORD = 4;
   */
  DOUBLE_WORD = 4,

  /**
   * @generated from enum value: TRIPLE_WORD = 5;
   */
  TRIPLE_WORD = 5,

  /**
   * @generated from enum value: QUADRUPLE_WORD = 6;
   */
  QUADRUPLE_WORD = 6,
}
// ...
export enum Direction {
  /**
   * @generated from enum value: UP = 0;
   */
  UP = 0,

  /**
   * @generated from enum value: DOWN = 1;
   */
  DOWN = 1,

  /**
   * @generated from enum value: LEFT = 2;
   */
  LEFT = 2,

  /**
   * @generated from enum value: RIGHT = 3;
   */
  RIGHT = 3,
}
// Retrieve enum metadata with: proto3.getEnumType(Direction)
proto3.util.setEnumType(Direction, "ScrabbleSharp.Direction", [
  { no: 0, name: "UP" },
  { no: 1, name: "DOWN" },
  { no: 2, name: "LEFT" },
  { no: 3, name: "RIGHT" },
]);

/**
 * @generated from message ScrabbleSharp.SolveRequest
 */
export class SolveRequest extends Message<SolveRequest> {
  /**
   * @generated from field: string board = 1;
   */
  board = "";

  /**
   * @generated from field: string rack = 2;
   */
  rack = "";

  constructor(data?: PartialMessage<SolveRequest>) {
    super();
    proto3.util.initPartial(data, this);
  }

  static readonly runtime: typeof proto3 = proto3;
  static readonly typeName = "ScrabbleSharp.SolveRequest";
  static readonly fields: FieldList = proto3.util.newFieldList(() => [
    { no: 1, name: "board", kind: "scalar", T: 9 /* ScalarType.STRING */ },
    { no: 2, name: "rack", kind: "scalar", T: 9 /* ScalarType.STRING */ },
  ]);

  static fromBinary(bytes: Uint8Array, options?: Partial<BinaryReadOptions>): SolveRequest {
    return new SolveRequest().fromBinary(bytes, options);
  }

  static fromJson(jsonValue: JsonValue, options?: Partial<JsonReadOptions>): SolveRequest {
    return new SolveRequest().fromJson(jsonValue, options);
  }

  static fromJsonString(jsonString: string, options?: Partial<JsonReadOptions>): SolveRequest {
    return new SolveRequest().fromJsonString(jsonString, options);
  }

  static equals(a: SolveRequest | PlainMessage<SolveRequest> | undefined, b: SolveRequest | PlainMessage<SolveRequest> | undefined): boolean {
    return proto3.util.equals(SolveRequest, a, b);
  }
}
// ...
export class SolveResponse extends Message<SolveResponse> {
  /**
   * @generated from field: repeated ScrabbleSharp.Move moves = 1;
   */
  moves: Move[] = [];

  constructor(data?: PartialMessage<SolveResponse>) {
    super();
    proto3.util.initPartial(data, this);
  }

  static readonly runtime: typeof proto3 = proto3;
  static readonly typeName = "ScrabbleSharp.SolveResponse";
  static readonly fields: FieldList = proto3.util.newFieldList(() => [
    { no: 1, name: "moves", kind: "message", T: Move, repeated: true },
  ]);

  static fromBinary(bytes: Uint8Array, options?: Partial<BinaryReadOptions>): SolveResponse {
    return new SolveResponse().fromBinary(bytes, options);
  }

  static fromJson(jsonValue: JsonValue, options?: Partial<JsonReadOptions>): SolveResponse {
    return new SolveResponse().fromJson(jsonValue, options);
  }

  static fromJsonString(jsonString: string, options?: Partial<JsonReadOptions>): SolveResponse {
    return new SolveResponse().fromJsonString(jsonString, options);
  }

  static equals(a: SolveResponse | PlainMessage<SolveResponse> | undefined, b: SolveResponse | PlainMessage<SolveResponse> | undefined): boolean {
    return proto3.util.equals(SolveResponse, a, b);
  }
}

/**
 * @generated from message ScrabbleSharp.LayoutRequest
 */
export class LayoutRequest extends Message<LayoutRequest> {
  /**
   * @generated from field: uint32 rows = 1;
   */
  rows = 0;

  /**
   * @generated from field: uint32 cols = 2;
   */
  cols = 0;

  constructor(data?: PartialMessage<LayoutRequest>) {
    super();
    proto3.util.initPartial(data, this);
  }

  static readonly runtime: typeof proto3 = proto3;
  static readonly typeName = "ScrabbleSharp.LayoutRequest";
  static readonly fields: FieldList = proto3.util.newFieldList(() => [
    { no: 1, name: "rows", kind: "scalar", T: 13 /* ScalarType.UINT32 */ },
    { no: 2, name: "cols", kind: "scalar", T: 13 /* ScalarType.UINT32 */ },
  ]);

  static fromBinary(bytes: Uint8Array, options?: Partial<BinaryReadOptions>): LayoutRequest {
    return new LayoutRequest().fromBinary(bytes, options);
  }

  static fromJson(jsonValue: JsonValue, options?: Partial<JsonReadOptions>): LayoutRequest {
    return new LayoutRequest().fromJson(jsonValue, options);
  }

  static fromJsonString(jsonString: string, options?: Partial<JsonReadOptions>): LayoutRequest {
    return new LayoutRequest().fromJsonString(jsonString, options);
  }

  static equals(a: LayoutRequest | PlainMessage<LayoutRequest> | undefined, b: LayoutRequest | PlainMessage<LayoutRequest> | undefined): boolean {
    return proto3.util.equals(LayoutRequest, a, b);
  }
}

/**
 * @generated from message ScrabbleSharp.LayoutResponse
 */
export class LayoutResponse extends Message<LayoutResponse> {
  /**
   * @generated from field: uint32 rows = 1;
   */
  rows = 0;

  /**
   * @generated from field: uint32 cols = 2;
   */
  cols = 0;

  /**
   * rows × cols, row-major
   *
   * @generated from field: repeated ScrabbleSharp.Multiplier multipliers = 3;
   */
  multipliers: Multiplier[] = [];

  /**
   * @generated from field: bool expandable = 4;
   */
  expandable = false;

  constructor(data?: PartialMessage<LayoutResponse>) {
    super();
    proto3.util.initPartial(data, this);
  }

  static readonly runtime: typeof proto3 = proto3;
  static readonly typeName = "ScrabbleSharp.LayoutResponse";
  static readonly fields: FieldList = proto3.util.newFieldList(() => [
    { no: 1, name: "rows", kind: "scalar", T: 13 /* ScalarType.UINT32 */ },
    { no: 2, name: "cols", kind: "scalar", T: 13 /* ScalarType.UINT32 */ },
    { no: 3, name: "multipliers", kind: "enum", T: proto3.getEnumType(Multiplier), repeated: true },
    { no: 4, name: "expandable", kind: "scalar", T: 8 /* ScalarType.BOOL */ },
  ]);

  static fromBinary(bytes: Uint8Array, options?: Partial<BinaryReadOptions>): LayoutResponse {
    return new LayoutResponse().fromBinary(bytes, options);
  }

  static fromJson(jsonValue: JsonValue, options?: Partial<JsonReadOptions>): LayoutResponse {
    return new LayoutResponse().fromJson(jsonValue, options);
  }

  static fromJsonString(jsonString: string, options?: Partial<JsonReadOptions>): LayoutResponse {
    return new LayoutResponse().fromJsonString(jsonString, options);
  }

  static equals(a: LayoutResponse | PlainMessage<LayoutResponse> | undefined, b: LayoutResponse | PlainMessage<LayoutResponse> | undefined): boolean {
    return proto3.util.equals(LayoutResponse, a, b);
  }
}

/**
 * @generated from message ScrabbleSharp.ExpandRequest
 */
export class ExpandRequest extends Message<ExpandRequest> {
  /**
   * @generated from field: ScrabbleSharp.Direction dir = 1;
   */
  dir = Direction.UP;

  constructor(data?: PartialMessage<ExpandRequest>) {
    super();
    proto3.util.initPartial(data, this);
  }

  static readonly runtime: typeof proto3 = proto3;
  static readonly typeName = "ScrabbleSharp.ExpandRequest";
  static readonly fields: FieldList = proto3.util.newFieldList(() => [
    { no: 1, name: "dir", kind: "enum", T: proto3.getEnumType(Direction) },
  ]);

  static fromBinary(bytes: Uint8Array, options?: Partial<BinaryReadOptions>): ExpandRequest {
    return new ExpandRequest().fromBinary(bytes, options);
  }

  static fromJson(jsonValue: JsonValue, options?: Partial<JsonReadOptions>): ExpandRequest {
    return new ExpandRequest().fromJson(jsonValue, options);
  }

  static fromJsonString(jsonString: string, options?: Partial<JsonReadOptions>): ExpandRequest {
    return new ExpandRequest().fromJsonString(jsonString, options);
  }

  static equals(a: ExpandRequest | PlainMessage<ExpandRequest> | undefined, b: ExpandRequest | PlainMessage<ExpandRequest> | undefined): boolean {
    return proto3.util.equals(ExpandRequest, a, b);
  }
}

/**
 * @generated from message ScrabbleSharp.ExpandDelta
 */
export class ExpandDelta extends Message<ExpandDelta> {
  /**
   * @generated from field: uint32 totalRows = 1;
   */
  totalRows = 0;

  /**
   * @generated from field: uint32 totalCols = 2;
   */
  totalCols = 0;

  /**
   * @generated from field: sint32 shiftRow = 3;
   */
  shiftRow = 0;

  /**
   * @generated from field: sint32 shiftCol = 4;
   */
  shiftCol = 0;

  /**
   * @generated from field: uint32 offsetRow = 5;
   */
  offsetRow = 0;

  /**
   * @generated from field: uint32 offsetCol = 6;
   */
  offsetCol = 0;

  /**
   * @generated from field: uint32 rows = 7;
   */
  rows = 0;

  /**
   * @generated from field: uint32 cols = 8;
   */
  cols = 0;

  /**
   * @generated from field: repeated ScrabbleSharp.Multiplier multipliers = 9;
   */
  multipliers: Multiplier[] = [];

  constructor(data?: PartialMessage<ExpandDelta>) {
    super();
    proto3.util.initPartial(data, this);
  }

  static readonly runtime: typeof proto3 = proto3;
  static readonly typeName = "ScrabbleSharp.ExpandDelta";
  static readonly fields: FieldList = proto3.util.newFieldList(() => [
    { no: 1, name: "totalRows", kind: "scalar", T: 13 /* ScalarType.UINT32 */ },
    { no: 2, name: "totalCols", kind: "scalar", T: 13 /* ScalarType.UINT32 */ },
    { no: 3, name: "shiftRow", kind: "scalar", T: 17 /* ScalarType.SINT32 */ },
    { no: 4, name: "shiftCol", kind: "scalar", T: 17 /* ScalarType.SINT32 */ },
    { no: 5, name: "offsetRow", kind: "scalar", T: 13 /* ScalarType.UINT32 */ },
    { no: 6, name: "offsetCol", kind: "scalar", T: 13 /* ScalarType.UINT32 */ },
    { no: 7, name: "rows", kind: "scalar", T: 13 /* ScalarType.UINT32 */ },
    { no: 8, name: "cols", kind: "scalar", T: 13 /* ScalarType.UINT32 */ },
    { no: 9, name: "multipliers", kind: "enum", T: proto3.getEnumType(Multiplier), repeated: true },
  ]);

  static fromBinary(bytes: Uint8Array, options?: Partial<BinaryReadOptions>): ExpandDelta {
    return new ExpandDelta().fromBinary(bytes, options);
  }

  static fromJson(jsonValue: JsonValue, options?: Partial<JsonReadOptions>): ExpandDelta {
    return new ExpandDelta().fromJson(jsonValue, options);
  }

  static fromJsonString(jsonString: string, options?: Partial<JsonReadOptions>): ExpandDelta {
    return new ExpandDelta().fromJsonString(jsonString, options);
  }

  static equals(a: ExpandDelta | PlainMessage<ExpandDelta> | undefined, b: ExpandDelta | PlainMessage<ExpandDelta> | undefined): boolean {
    return proto3.util.equals(ExpandDelta, a, b);
  }
}